fix(types): clear mypy debt so the 0.8.0 release passes the strict mypy gate
The 0.8.0 source carried 320 mypy errors under the public CI gate (the monorepo
CI did not run mypy, so the type debt only surfaced here). Cleared type-only
(no behavior change; monorepo full suite 2020 passed): annotations, Optional for
None-defaults, None-narrowing, any->Any, container element types, cast() at DB
boundaries, and restored otel-fallback type:ignore on Status/StatusCode.

+++ phionyx_core/memory/rag_cache.py
@@ -347,7 +347,7 @@ class RAGCache:
 
         return len(expired_keys)
 
-    def get_metrics(self) -> Dict[str, int]:
+    def get_metrics(self) -> Dict[str, float]:
         """
         Get cache metrics.
